requirements.txt
flask>=3.0
sqlalchemy>=2.0
pymysql>=1.1
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.0
apscheduler>=3.10
openpyxl>=3.1
python-dotenv>=1.0
